Split DbConnectionStringCommon classes into separate classes (will deal with utils in a bit)
