+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/Common/DbConnectionStringCommon.cs
@@ -950,215 +950,9 @@ namespace Microsoft.Data.Common
         #endregion
     }
 
-    internal static class DbConnectionStringDefaults
-    {
-        internal const ApplicationIntent ApplicationIntent = Microsoft.Data.SqlClient.ApplicationIntent.ReadWrite;
-        internal const string ApplicationName =
-#if NETFRAMEWORK
-            "Framework Microsoft SqlClient Data Provider";
-#else
-            "Core Microsoft SqlClient Data Provider";
-#endif
-        internal const string AttachDBFilename = "";
-        internal const int CommandTimeout = 30;
-        internal const int ConnectTimeout = 15;
-        internal const bool ContextConnection = false;
+    
 
-#if NETFRAMEWORK
-        internal const bool ConnectionReset = true;
-        internal static readonly bool TransparentNetworkIPResolution = !LocalAppContextSwitches.DisableTNIRByDefault;
-        internal const string NetworkLibrary = "";
-#endif
-        internal const string CurrentLanguage = "";
-        internal const string DataSource = "";
-        internal static readonly SqlConnectionEncryptOption Encrypt = SqlConnectionEncryptOption.Mandatory;
-        internal const string HostNameInCertificate = "";
-        internal const string ServerCertificate = "";
-        internal const bool Enlist = true;
-        internal const string FailoverPartner = "";
-        internal const string InitialCatalog = "";
-        internal const bool IntegratedSecurity = false;
-        internal const int LoadBalanceTimeout = 0; // default of 0 means don't use
-        internal const bool MultipleActiveResultSets = false;
-        internal const bool MultiSubnetFailover = false;
-        internal const int MaxPoolSize = 100;
-        internal const int MinPoolSize = 0;
-        internal const int PacketSize = 8000;
-        internal const string Password = "";
-        internal const bool PersistSecurityInfo = false;
-        internal const bool Pooling = true;
-        internal const bool TrustServerCertificate = false;
-        internal const string TypeSystemVersion = "Latest";
-        internal const string UserID = "";
-        internal const bool UserInstance = false;
-        internal const bool Replication = false;
-        internal const string WorkstationID = "";
-        internal const string TransactionBinding = "Implicit Unbind";
-        internal const int ConnectRetryCount = 1;
-        internal const int ConnectRetryInterval = 10;
-        internal static readonly SqlAuthenticationMethod Authentication = SqlAuthenticationMethod.NotSpecified;
-        internal const SqlConnectionColumnEncryptionSetting ColumnEncryptionSetting = SqlConnectionColumnEncryptionSetting.Disabled;
-        internal const string EnclaveAttestationUrl = "";
-        internal const SqlConnectionAttestationProtocol AttestationProtocol = SqlConnectionAttestationProtocol.NotSpecified;
-        internal const SqlConnectionIPAddressPreference IPAddressPreference = SqlConnectionIPAddressPreference.IPv4First;
-        internal const PoolBlockingPeriod PoolBlockingPeriod = SqlClient.PoolBlockingPeriod.Auto;
-        internal const string ServerSPN = "";
-        internal const string FailoverPartnerSPN = "";
-    }
+    
 
-    internal static class DbConnectionStringKeywords
-    {
-#if NETFRAMEWORK
-        // Odbc
-        internal const string Driver = "Driver";
-        internal const string Dsn = "Dsn";
-        internal const string FileDsn = "FileDsn";
-        internal const string SaveFile = "SaveFile";
-
-        // OleDb
-        internal const string FileName = "File Name";
-        internal const string OleDbServices = "OLE DB Services";
-        internal const string Provider = "Provider";
-
-        // OracleClient
-        internal const string Unicode = "Unicode";
-        internal const string OmitOracleConnectionName = "Omit Oracle Connection Name";
-#endif
-        // SqlClient
-        internal const string ApplicationIntent = "Application Intent";
-        internal const string ApplicationName = "Application Name";
-        internal const string AttachDBFilename = "AttachDbFilename";
-        internal const string ConnectTimeout = "Connect Timeout";
-        internal const string CommandTimeout = "Command Timeout";
-        internal const string ConnectionReset = "Connection Reset";
-        internal const string ContextConnection = "Context Connection";
-        internal const string CurrentLanguage = "Current Language";
-        internal const string Encrypt = "Encrypt";
-        internal const string HostNameInCertificate = "Host Name In Certificate";
-        internal const string ServerCertificate = "Server Certificate";
-        internal const string FailoverPartner = "Failover Partner";
-        internal const string InitialCatalog = "Initial Catalog";
-        internal const string MultipleActiveResultSets = "Multiple Active Result Sets";
-        internal const string MultiSubnetFailover = "Multi Subnet Failover";
-        internal const string NetworkLibrary = "Network Library";
-        internal const string PacketSize = "Packet Size";
-        internal const string Replication = "Replication";
-        internal const string TransactionBinding = "Transaction Binding";
-        internal const string TrustServerCertificate = "Trust Server Certificate";
-        internal const string TypeSystemVersion = "Type System Version";
-        internal const string UserInstance = "User Instance";
-        internal const string WorkstationID = "Workstation ID";
-        internal const string ConnectRetryCount = "Connect Retry Count";
-        internal const string ConnectRetryInterval = "Connect Retry Interval";
-        internal const string Authentication = "Authentication";
-        internal const string ColumnEncryptionSetting = "Column Encryption Setting";
-        internal const string EnclaveAttestationUrl = "Enclave Attestation Url";
-        internal const string AttestationProtocol = "Attestation Protocol";
-        internal const string IPAddressPreference = "IP Address Preference";
-        internal const string ServerSPN = "Server SPN";
-        internal const string FailoverPartnerSPN = "Failover Partner SPN";
-        internal const string TransparentNetworkIPResolution = "Transparent Network IP Resolution";
-
-        // common keywords (OleDb, OracleClient, SqlClient)
-        internal const string DataSource = "Data Source";
-        internal const string IntegratedSecurity = "Integrated Security";
-        internal const string Password = "Password";
-        internal const string PersistSecurityInfo = "Persist Security Info";
-        internal const string UserID = "User ID";
-
-        // managed pooling (OracleClient, SqlClient)
-        internal const string Enlist = "Enlist";
-        internal const string LoadBalanceTimeout = "Load Balance Timeout";
-        internal const string MaxPoolSize = "Max Pool Size";
-        internal const string Pooling = "Pooling";
-        internal const string MinPoolSize = "Min Pool Size";
-        internal const string PoolBlockingPeriod = "Pool Blocking Period";
-    }
-
-    internal static class DbConnectionStringSynonyms
-    {
-        //internal const string TransparentNetworkIPResolution = TRANSPARENTNETWORKIPRESOLUTION;
-        internal const string TRANSPARENTNETWORKIPRESOLUTION = "transparentnetworkipresolution";
-
-        //internal const string ApplicationName        = APP;
-        internal const string APP = "app";
-
-        // internal const string IPAddressPreference = IPADDRESSPREFERENCE;
-        internal const string IPADDRESSPREFERENCE = "ipaddresspreference";
-
-        //internal const string ApplicationIntent    = APPLICATIONINTENT;
-        internal const string APPLICATIONINTENT = "applicationintent";
-
-        //internal const string AttachDBFilename       = EXTENDEDPROPERTIES+","+INITIALFILENAME;
-        internal const string EXTENDEDPROPERTIES = "extended properties";
-        internal const string INITIALFILENAME = "initial file name";
-
-        // internal const string HostNameInCertificate        = HOSTNAMEINCERTIFICATE;
-        internal const string HOSTNAMEINCERTIFICATE = "hostnameincertificate";
-
-        // internal const string ServerCertificate        = SERVERCERTIFICATE;
-        internal const string SERVERCERTIFICATE = "servercertificate";
-
-        //internal const string ConnectTimeout         = CONNECTIONTIMEOUT+","+TIMEOUT;
-        internal const string CONNECTIONTIMEOUT = "connection timeout";
-        internal const string TIMEOUT = "timeout";
-
-        //internal const string ConnectRetryCount = CONNECTRETRYCOUNT;
-        internal const string CONNECTRETRYCOUNT = "connectretrycount";
-
-        //internal const string ConnectRetryInterval = CONNECTRETRYINTERVAL;
-        internal const string CONNECTRETRYINTERVAL = "connectretryinterval";
-
-        //internal const string CurrentLanguage        = LANGUAGE;
-        internal const string LANGUAGE = "language";
-
-        //internal const string OraDataSource          = SERVER;
-        //internal const string SqlDataSource          = ADDR+","+ADDRESS+","+SERVER+","+NETWORKADDRESS;
-        internal const string ADDR = "addr";
-        internal const string ADDRESS = "address";
-        internal const string SERVER = "server";
-        internal const string NETWORKADDRESS = "network address";
-
-        //internal const string InitialCatalog         = DATABASE;
-        internal const string DATABASE = "database";
-
-        //internal const string IntegratedSecurity     = TRUSTEDCONNECTION;
-        internal const string TRUSTEDCONNECTION = "trusted_connection"; // underscore introduced in everett
-
-        //internal const string LoadBalanceTimeout     = ConnectionLifetime;
-        internal const string ConnectionLifetime = "connection lifetime";
-
-        //internal const string MultipleActiveResultSets    = MULTIPLEACTIVERESULTSETS;
-        internal const string MULTIPLEACTIVERESULTSETS = "multipleactiveresultsets";
-
-        //internal const string MultiSubnetFailover = MULTISUBNETFAILOVER;
-        internal const string MULTISUBNETFAILOVER = "multisubnetfailover";
-
-        //internal const string NetworkLibrary         = NET+","+NETWORK;
-        internal const string NET = "net";
-        internal const string NETWORK = "network";
-
-        //internal const string PoolBlockingPeriod = POOLBLOCKINGPERIOD;
-        internal const string POOLBLOCKINGPERIOD = "poolblockingperiod";
-
-        //internal const string Password               = Pwd;
-        internal const string Pwd = "pwd";
-
-        //internal const string PersistSecurityInfo    = PERSISTSECURITYINFO;
-        internal const string PERSISTSECURITYINFO = "persistsecurityinfo";
-
-        //internal const string TrustServerCertificate = TRUSTSERVERCERTIFICATE;
-        internal const string TRUSTSERVERCERTIFICATE = "trustservercertificate";
-
-        //internal const string UserID                 = UID+","+User;
-        internal const string UID = "uid";
-        internal const string User = "user";
-
-        //internal const string WorkstationID          = WSID;
-        internal const string WSID = "wsid";
-
-        //internal const string server SPNs
-        internal const string ServerSPN = "ServerSPN";
-        internal const string FailoverPartnerSPN = "FailoverPartnerSPN";
-    }
+    
 }
